requirements.txt
adversarial_robustness_toolbox==1.20.1
matplotlib==3.10.5
numpy==2.3.2
pandas==1.5.3
Pillow==11.3.0
scikit_learn==1.7.1
seaborn==0.13.2
torch==2.6.0a0+ecf3bae40a.nv25.1
torchmetrics==1.8.1
torchvision==0.20.0a0
